+++ packages/business/const/src/branding.ts
@@ -2,26 +2,26 @@
 // if you want to use it in the commercial usage
 // please contact us for more information: hello@lobehub.com
 
-export const LOBE_CHAT_CLOUD = 'LobeHub Cloud';
+export const LOBE_CHAT_CLOUD = 'mAI Cloud';
 
-export const BRANDING_NAME = 'LobeHub';
-export const BRANDING_LOGO_URL = '';
+export const BRANDING_NAME = 'mAI';
+export const BRANDING_LOGO_URL = '/avatars/may.PNG';
 
-export const ORG_NAME = 'LobeHub';
+export const ORG_NAME = 'mAI';
 
 export const BRANDING_URL = {
-  help: undefined,
-  privacy: undefined,
+  help: 'https://mprojects.odoo.com',
+  privacy: 'https://mprojects.odoo.com',
   subscription: 'https://app.lobehub.com/settings/plans',
-  support: undefined,
-  terms: undefined,
+  support: 'https://mprojects.odoo.com',
+  terms: 'https://mprojects.odoo.com',
 };
 
 export const SOCIAL_URL = {
-  discord: 'https://discord.gg/AYFPHvv2jT',
-  github: 'https://github.com/lobehub',
-  medium: 'https://medium.com/@lobehub',
-  x: 'https://x.com/lobehub',
+  discord: 'https://discord.com/invite/fV7zwdGPpY',
+  github: 'https://github.com/mDevsLabs/mAI',
+  medium: 'https://medium.com/mdevslabs',
+  x: 'https://x.com/mdevslabs',
   youtube: 'https://www.youtube.com/@lobehub',
 };