+++ packages/const/src/url.ts
@@ -6,27 +6,27 @@ export const OFFICIAL_DOMAIN = 'lobehub.com';
 
 export const OG_URL = '/og/og.webp?v=1';
 
-export const GITHUB = 'https://github.com/lobehub/lobe-chat';
+export const GITHUB = 'https://github.com/mDevsLabs/mAI';
 export const GITHUB_ISSUES = urlJoin(GITHUB, 'issues/new/choose');
-export const CHANGELOG = 'https://lobehub.com/changelog';
+export const CHANGELOG = 'https://mprojects.odoo.com';
 
-export const DOCUMENTS = urlJoin(OFFICIAL_SITE, '/docs');
-export const USAGE_DOCUMENTS = urlJoin(DOCUMENTS, '/usage');
-export const SELF_HOSTING_DOCUMENTS = urlJoin(DOCUMENTS, '/self-hosting');
-export const DATABASE_SELF_HOSTING_URL = urlJoin(SELF_HOSTING_DOCUMENTS, '/server-database');
+export const DOCUMENTS = 'https://mprojects.odoo.com';
+export const USAGE_DOCUMENTS = 'https://mprojects.odoo.com';
+export const SELF_HOSTING_DOCUMENTS = 'https://mprojects.odoo.com';
+export const DATABASE_SELF_HOSTING_URL = 'https://mprojects.odoo.com';
 
 // use this for the link
-export const DOCUMENTS_REFER_URL = `${DOCUMENTS}?utm_source=chat_preview`;
+export const DOCUMENTS_REFER_URL = 'https://mprojects.odoo.com';
 
-export const WIKI_PLUGIN_GUIDE = urlJoin(USAGE_DOCUMENTS, '/plugins/development');
-export const MANUAL_UPGRADE_URL = urlJoin(SELF_HOSTING_DOCUMENTS, '/advanced/upstream-sync');
+export const WIKI_PLUGIN_GUIDE = 'https://mprojects.odoo.com';
+export const MANUAL_UPGRADE_URL = 'https://mprojects.odoo.com';
 
-export const BLOG = urlJoin(OFFICIAL_SITE, 'blog');
+export const BLOG = 'https://mprojects.odoo.com';
 
-export const ABOUT = OFFICIAL_SITE;
-export const FEEDBACK = 'https://github.com/lobehub/lobe-chat/issues/new/choose';
-export const PRIVACY_URL = urlJoin(OFFICIAL_SITE, '/privacy');
-export const TERMS_URL = urlJoin(OFFICIAL_SITE, '/terms');
+export const ABOUT = 'https://mprojects.odoo.com';
+export const FEEDBACK = 'https://mprojects.odoo.com';
+export const PRIVACY_URL = 'https://mprojects.odoo.com';
+export const TERMS_URL = 'https://mprojects.odoo.com';
 
 export const PLUGINS_INDEX_URL = 'https://chat-plugins.lobehub.com';
 
@@ -70,8 +70,8 @@ export const RELEASES_URL = urlJoin(GITHUB, 'releases');
 export const mailTo = (email: string) => `mailto:${email}`;
 
 export const AES_GCM_URL = 'https://datatracker.ietf.org/doc/html/draft-ietf-avt-srtp-aes-gcm-01';
-export const BASE_PROVIDER_DOC_URL = 'https://lobehub.com/docs/usage/providers';
-export const CHANGELOG_URL = urlJoin(OFFICIAL_SITE, 'changelog');
+export const BASE_PROVIDER_DOC_URL = 'https://mprojects.odoo.com';
+export const CHANGELOG_URL = 'https://mprojects.odoo.com';
 
 export const DOWNLOAD_URL = {
   android: 'https://play.google.com/store/apps/details?id=com.lobehub.app',